added function.php
