+++ index.php
@@ -1,22 +1,6 @@
 <?php
 require_once 'helpers.php';
-
-function date_calculating($date) {
-    if($date === 'null') { return '';}
-
-    $current_date = time();
-    $hour = 3600;
-
-    $transfered_date = strtotime($date);
-
-    $difference = floor(($transfered_date - $current_date) / $hour);
-    $difference_day = floor($difference / 24);
-    if($difference <= 24) {
-        echo '<small class="badge badge-danger"><i class="far fa-clock"></i>'.str_pad($difference, strlen($difference) + 11 ,"Годин:",STR_PAD_LEFT).'</small>';
-    } else {echo '<small class="badge badge-success"><i class="far fa-clock"></i>'.str_pad($difference_day, strlen($difference_day) +9,"Днів:",STR_PAD_LEFT).'</small>';}
-}
-
-
+require_once 'functions.php';
 $user_name = 'Володимир';
 $user_image = 'static/img/user2-160x160.jpg';
 $category = ['Вхідні', 'Навчання', 'Робота', 'Домашні справи', 'Авто'];
@@ -56,21 +40,9 @@ $info = [
     ],
 ];
 
-function taskSum( array $array, string $projectName): int
-{
-    $number = 0;
-    foreach ($array as $el) {
-        if ($el['category'] === $projectName) {
-            $number += 1;
-        }
-    }
-    return $number;
-}
-
 $kanban_template = renderTemplate(
      'kanban.php',
-     ['info' => $info,
-         ]);
+     ['info' => $info]);
 
 $main_content =renderTemplate('main.php', [
     'wrapper_content'=> $kanban_template,
